+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Release Notes
 
-## Version 3.0 - 03/16/2020
+## Version 1.0.1 - 06/26/2020
+* Added cache command to list and clean the containerd docker images cache
+* Generated TAR only contains the extracted images for the chart, instead of all the images contained in the cache
+
+## Version 1.0 - 06/20/2020
 * First delivery on Github.